* added some more JsonSerialize Informations * added first Finder method and * added persist methods in SyncBU

+++ src/main/java/org/synyx/sytim/spring/boot/domain/TimeEntry.java
@@ -17,7 +17,7 @@ import org.springframework.data.neo4j.annotation.NodeEntity;
 public class TimeEntry {
 
     @GraphId
-    public int graphId;
+    public Long graphId;
 
     public String desc;
